fix: CSS 코드 수정
