+++ front-end/src/components/myPage/Myinfo.jsx
@@ -5,13 +5,13 @@ const MyInfo = ({ myInfo }) => {
   return (
     <div>
       <h1>내 정보</h1>
-      <div style={{ display: "flex", justifyContent: "space-between" }}>
-        <div style={{ marginBottom: "100px" }}>
+      <div>
+        <div>
           <p>회원번호: {myInfo.id}</p>
           <p>닉네임: {myInfo.nick}</p>
           <p>이메일: {myInfo.email}</p>
         </div>
-        <div style={{ marginLeft: "auto" }}>
+        <div>
           <MonsterBall />
         </div>
       </div>
