+++ front-end/src/pages/MyPage.jsx
@@ -1,11 +1,10 @@
-import MyMain from "/@/components/myPage/MyMain";
+import MyMain from "../components/myPage/MyMain";
 
 const MyPage = () => {
   return (
-    <div>
-      <h1>마이 페이지</h1>
+    <>
       <MyMain />
-    </div>
+    </>
   );
 };
